requirements.txt
fastapi==0.115.5
spacy~=3.8.2
uvicorn==0.22.0
python-dotenv==1.0.1
requests~=2.32.3
transformers~=4.46.3
torch~=2.5.1
torchaudio==2.5.0
torchvision==0.20.0
numpy==1.26.3
nltk~=3.9.1
pandas==2.1.0
openpyxl==3.1.2
datasets==3.0.0
huggingface-hub==0.23.3

pydantic~=2.10.1
pip~=24.3.1
wheel~=0.45.1
thinc~=8.3.2
attrs~=24.2.0
pillow~=11.0.0
filelock~=3.16.1